Update the documentation and template for the generator
Users should not have to set their own AUTHIC_CLIENT_FULL_URL as it
is just the subdomain, on authic.com. It is still useful to be able
to override this constant as applications setup to test staging and
developement servers will have different urls to the default
authic.com. The end user has no need to know this.